fix(tests): stop conftest from being imported a second time
`tests/conftest.py` puts `tests/` on sys.path, so `from conftest import ...`
in 13 modules resolved to a *second*, independent module object alongside
the one pytest imports as `tests.conftest`. Both executed the module body,
and the later one — the plain `conftest`, imported during collection —
won the assignment that matters:

    sys.modules["decky"] = mock_decky

Fixtures are registered only from pytest's copy, so `_reset_decky_mock_paths`
refreshed a mock that nothing observed, while production code reading
`decky.DECKY_PLUGIN_SETTINGS_DIR` kept seeing the shadow's import-time
directory for the whole session. The per-test isolation the fixture exists
to provide was therefore silently absent — and only in runs that collected
one of the 13 importers, which makes the behaviour depend on the test
selection:

    tests/adapters alone            -> fresh dir per test
    the same file + tests/services  -> one shared dir for every test

Nothing depended on it yet: the only test reaching `Plugin._main()` patches
`main.bootstrap`, so the paths are never used for real I/O. It was a trap
waiting for the first persistence test, not an active failure.

The shared constructors move to `tests/_factories.py`, matching the helper
modules the suite already uses (`tests/contract/_harness.py`,
`tests/services/saves/_helpers.py`). Nothing imports `conftest` now, so a
single module object survives and the fixture reaches the live mock.

Switching the import mode to `prepend` would fix the root cause too, but is
not available here: 17 test basenames repeat across directories and only 6
of them sit in packages.
