+++ tests/conftest.py
@@ -4,7 +4,6 @@ import os
 import shutil
 import sys
 import tempfile
-from typing import Any
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
@@ -42,10 +41,11 @@ def _cleanup_process_temp_dirs() -> None:
     """Drop the import-time temp dirs — ``mkdtemp`` never cleans up after itself.
 
     Deliberately ``atexit`` rather than a session-scoped fixture: a fixture
-    only tears down when at least one test is collected, and this module is
-    imported twice per run (pytest's own copy plus the plain ``conftest``
-    that `tests/` modules import directly). Both copies register here, so
-    both sets of directories are released.
+    only tears down once at least one test is collected, so a run that
+    collects nothing — pointing pytest at a directory that holds only
+    fixtures or vectors — would leak all three. atexit also stays correct
+    if this module is ever executed more than once, since each execution
+    releases exactly what it created.
     """
     for path in (_process_settings_dir, _process_runtime_dir, _process_log_dir):
         shutil.rmtree(path, ignore_errors=True)
@@ -62,74 +62,6 @@ mock_decky.logger = logging.getLogger("test_romm")
 mock_decky.emit = AsyncMock()
 
 sys.modules["decky"] = mock_decky
-
-
-def _no_retry(fn, *a, **kw):
-    """Pass-through Retry side_effect: invoke the wrapped callable once, no backoff."""
-    return fn(*a, **kw)
-
-
-def _make_retry():
-    """Build a Retry ``MagicMock`` that runs ``with_retry`` callables exactly once
-    and reports every exception as non-retryable. Used everywhere services
-    take a ``Retry`` Protocol injection in tests."""
-    retry = MagicMock()
-    retry.with_retry.side_effect = _no_retry
-    retry.is_retryable.return_value = False
-    return retry
-
-
-def _make_testable_plugin():
-    """Return a TestablePlugin instance with test-only attributes declared.
-
-    Pre-populates ``_migration_service`` with a non-pending MagicMock so the
-    ``@migration_blocked`` decorator passes through (it requires the service and
-    raises RuntimeError if it is unwired) in tests that don't otherwise wire
-    migration state. Tests that exercise the block can override
-    ``is_retrodeck_migration_pending`` per-test.
-
-    Also pre-wires a no-op ``_debug_logger`` so any service that consumes
-    ``Plugin._log_debug`` (which forwards through ``_debug_logger``) works
-    out of the box. Tests that want to assert on debug-log behaviour can
-    override ``_debug_logger`` after construction (e.g. with the real
-    ``SettingsAwareDebugLogger`` bound to a settings dict they control).
-    """
-    # Import here to ensure decky mock is already installed
-    from main import Plugin
-
-    class TestablePlugin(Plugin):
-        """Plugin subclass that declares test-only attributes for type safety.
-
-        Genuinely test-fixture-only attributes live here: ``_fake_api``,
-        ``_resolve_system``, ``_save_settings``, plus the Unit-of-Work
-        handles tests seed and assert against (``_uow``, ``_uow_factory``)
-        and the per-test ``_tmp_path`` scratch dir. Test-fixture handles
-        shared with production wiring (``_state``, ``_http_adapter``, ...)
-        are declared on ``Plugin`` itself as ``Any``-typed annotation slots
-        so test-only construction paths type-check uniformly.
-        ``_save_settings`` is a test-only handle for the settings dict tests
-        thread into ``SaveService`` / ``PlaytimeService``; production threads
-        its settings store as ``self.settings``, never under this name.
-        """
-
-        _fake_api: Any
-        _resolve_system: Any
-        _save_settings: Any
-        _uow: Any
-        _uow_factory: Any
-        _tmp_path: Any
-        _core_info: Any
-        _platform_core_reader: Any
-        _active_core: Any
-        _m3u_supported: Any
-        _renderer_rss: Any
-        _renderer_gc: Any
-
-    instance = TestablePlugin()
-    instance._migration_service = MagicMock()
-    instance._migration_service.is_retrodeck_migration_pending.return_value = False
-    instance._debug_logger = lambda msg: None
-    return instance
 
 
 @pytest.fixture
